requirements.txt
ansible>=6.0.0
ansible-core>=2.14.0
netmiko>=4.0.0
paramiko>=2.9.0
pyyaml>=6.0
jinja2>=3.0.0
pyserial>=3.5
scp>=0.14.0
textfsm>=1.1.0
ntc-templates>=3.0.0